pyproject.toml
[project]
name = "messengerapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "argon2-cffi>=25.1.0",
    "fastapi[standard]>=0.124.4",
    "sqlalchemy>=2.0.45",
]

[dependency-groups]
dev = [
    "ruff>=0.14.9",
]
